[BE] Refactor: 보관함 아티스트 서비스 분리 및 적용
