+++ backend/src/route/library/artist/controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
-import User from '../../../entities/User';
-import Artist from '../../../entities/Artist';
+import { IJwtPayload } from '../../../middlewares/auth';
+import * as libraryArtistService from '../../../services/library/artist';
 
 const getArtistsByUserId = async (
   req: Request,
@@ -8,12 +8,9 @@ const getArtistsByUserId = async (
   next: NextFunction,
 ): Promise<void> => {
   try {
-    const { id: userId } = req.user as User;
-    const artistList = await Artist.findByUserId(userId);
-    res.status(200).json({
-      success: true,
-      data: artistList,
-    });
+    const { id: userId } = req.user as IJwtPayload;
+    const artistList = await libraryArtistService.getArtistsByUserId(userId);
+    res.status(200).json({ success: true, data: artistList });
   } catch (err) {
     console.log(err);
     next(err);
@@ -22,16 +19,12 @@ const getArtistsByUserId = async (
 
 const addArtist = async (req: Request, res: Response, next: NextFunction): Promise<any> => {
   try {
-    const { id: userId } = req.user as User;
+    const { id: userId } = req.user as IJwtPayload;
     const { artistId } = req.body;
     if (!artistId) return res.status(400).json({ message: 'Parameter Error: artistId' });
 
-    const user = (await User.findOne(userId, { relations: ['artists'] })) as User;
-    const artist = (await Artist.findOne(artistId)) as Artist;
-    if (!artist) return res.status(404).json({ message: 'Artist Not Found' });
-
-    user.artists.push(artist);
-    await user.save();
+    const isSuccess = libraryArtistService.addArtist(userId, artistId);
+    if (!isSuccess) return res.status(404).json({ message: 'Artist Not Found' });
     return res.status(204).end();
   } catch (err) {
     console.log(err);
@@ -41,16 +34,12 @@ const addArtist = async (req: Request, res: Response, next: NextFunction): Promi
 
 const deleteArtist = async (req: Request, res: Response, next: NextFunction): Promise<any> => {
   try {
-    const { id: userId } = req.user as User;
+    const { id: userId } = req.user as IJwtPayload;
     const { artistId } = req.params;
     if (!artistId) return res.status(400).json({ message: 'Parameter Error: artistId' });
 
-    const user = (await User.findOne(userId, { relations: ['artists'] })) as User;
-    const artistToRemove = (await Artist.findOne(artistId)) as Artist;
-    if (!artistToRemove) return res.status(404).json({ message: 'Artist Not Found' });
-
-    user.artists = user.artists.filter(artist => artist.id !== artistToRemove.id);
-    await user.save();
+    const isSuccess = libraryArtistService.deleteArtist(userId, parseInt(artistId, 10));
+    if (!isSuccess) return res.status(404).json({ message: 'Artist Not Found' });
     return res.status(204).end();
   } catch (err) {
     console.log(err);
